### 2019-01-15 7a49142
Biddy 1.8.2 (SVN revision 539)
