+++ paper.bib
@@ -1,21 +1,15 @@
 @online{Biddy,
   author = {Robert Meolic},
   title = {Biddy's homepage},
-  url = {http://biddy.meolic.com/},
-  urldate = {2019-01-15}
-}
-
-@online{BiddyUserManual,
-  author = {Robert Meolic},
-  title = {Biddy 1.8.2: User Manual},
   year = 2018,
-  url = {http://biddy.meolic.com/doxygen/html/},
+  url = {http://biddy.meolic.com/},
   urldate = {2019-01-15}
 }
 
 @online{BddEncyclopedia,
   author = {Robert Meolic},
-  title = {BDD Encyclopedia},
+  title = {{BDD} Encyclopedia},
+  year = 2018,
   url = {http://svn.savannah.nongnu.org/viewvc/*checkout*/biddy/bddscout/ENCYCLOPEDIA/bddencyclopedia.html},
   urldate = {2019-01-15}
 }
@@ -27,44 +21,44 @@
   doi = {10.1109/TC.1986.1676819}
 }
 
-@book{Knuth2009,
+@article{Knuth2009,
   author = {D. E. Knuth},
-  title = {The Art of Computer Programming, Volume 4, Fascicle 1: Bitwise Tricks & Techniques; Binary Decision Diagrams. Addison-Wesley Professional, 2009},
+  title = {The Art of Computer Programming, Volume 4, Fascicle 1: Bitwise Tricks and Techniques; Binary Decision Diagrams. Addison-Wesley Professional, 2009},
   year = 2009
 }
 
 @article{Minato2013,
   author = {Shin-ichi Minato},
-  title = {Techniques of BDD/ZDD: Brief History and Recent Activity. IEICE Trans. on Information and Systems, E96D(7), 1419–1429, 2013.},
+  title = {Techniques of {BDD}/{ZDD}: Brief History and Recent Activity. IEICE Trans. on Information and Systems, E96D(7), 1419–1429, 2013},
   year = 2013,
   doi = {10.1587/transinf.E96.D.1419}
 }
 
 @article{Meolic2012,
   author = {Robert Meolic},
-  title = {Biddy - a Multi-platform Academic BDD Package. Journal of Software, 7(6), 1358–1366, 2012.},
+  title = {Biddy - a Multi-platform Academic {BDD} Package. Journal of Software, 7(6), 1358–1366, 2012.},
   year = 2012,
   doi = {10.4304/jsw.7.6.1358-1366}
 }
 
 @article{Meolic2016,
   author = {Robert Meolic},
-  title = {Implementation aspects of a BDD package supporting general decision diagrams. Tech. rep., University of Maribor, 2016},
+  title = {Implementation aspects of a {BDD} package supporting general decision diagrams. Tech. rep., University of Maribor, 2016},
   year = 2016,
   url = {https://dk.um.si/IzpisGradiva.php?id=68831&lang=eng},
   urldate = {2019-01-15}
 }
 
 @article{Meolic2017,
-  author = {Tom van Dijk, Robert Wille, Robert Meolic},
-  title = {Tagged BDDs: combining reduction rules from different decision diagram types. In Proc. FMCAD 2017, pp. 108-115, 2017.},
+  author = {Tom van Dijk and Robert Wille, and Robert Meolic},
+  title = {Tagged {BDDs}: combining reduction rules from different decision diagram types. In Proc. FMCAD 2017, pp. 108-115, 2017},
   year = 2017,
   doi = {10.23919/FMCAD.2017.8102248}
 }
 
 @article{Meolic2018,
-  author = {Robert Meolic, Zmago Brezočnik},
[…]
