+++ paper.bib
@@ -1,22 +1,6 @@
-@online{Biddy,
-  author = {Robert Meolic},
-  title = {Biddy's homepage},
-  year = 2018,
-  url = {http://biddy.meolic.com/},
-  urldate = {2019-01-15}
-}
-
-@online{BddEncyclopedia,
-  author = {Robert Meolic},
-  title = {{BDD} Encyclopedia},
-  year = 2018,
-  url = {http://svn.savannah.nongnu.org/viewvc/*checkout*/biddy/bddscout/ENCYCLOPEDIA/bddencyclopedia.html},
-  urldate = {2019-01-15}
-}
-
 @article{Bryant1986,
   author = {R. E. Bryant},
-  title = {Graph-Based Algorithms for Boolean Function Manipulation. IEEE Transactions on Computers, 35(8), 677–691, 1986.},
+  title = {Graph-Based Algorithms for Boolean Function Manipulation. IEEE Transactions on Computers, 35(8), 677–691, 1986},
   year = 1986,
   doi = {10.1109/TC.1986.1676819}
 }
@@ -27,38 +11,54 @@
   year = 2009
 }
 
-@article{Minato2013,
-  author = {Shin-ichi Minato},
-  title = {Techniques of {BDD}/{ZDD}: Brief History and Recent Activity. IEICE Trans. on Information and Systems, E96D(7), 1419–1429, 2013},
-  year = 2013,
-  doi = {10.1587/transinf.E96.D.1419}
-}
-
 @article{Meolic2012,
-  author = {Robert Meolic},
-  title = {Biddy - a Multi-platform Academic {BDD} Package. Journal of Software, 7(6), 1358–1366, 2012.},
+  author = {R. Meolic},
+  title = {Biddy - a Multi-platform Academic {BDD} Package. Journal of Software, 7(6), 1358–1366, 2012},
   year = 2012,
   doi = {10.4304/jsw.7.6.1358-1366}
 }
 
 @article{Meolic2016,
-  author = {Robert Meolic},
-  title = {Implementation aspects of a {BDD} package supporting general decision diagrams. Tech. rep., University of Maribor, 2016},
+  author = {R. Meolic},
+  title = {Implementation aspects of a {BDD} package supporting general decision diagrams. {U}niversity of {M}aribor, Tech. rep., 2016},
   year = 2016,
   url = {https://dk.um.si/IzpisGradiva.php?id=68831&lang=eng},
   urldate = {2019-01-15}
 }
 
 @article{Meolic2017,
-  author = {Tom van Dijk and Robert Wille, and Robert Meolic},
+  author = {T. van Dijk and R. Wille and R. Meolic},
   title = {Tagged {BDDs}: combining reduction rules from different decision diagram types. In Proc. FMCAD 2017, pp. 108-115, 2017},
   year = 2017,
   doi = {10.23919/FMCAD.2017.8102248}
 }
 
 @article{Meolic2018,
-  author = {Robert Meolic and Zmago Brezočnik},
+  author = {R. Meolic and Z. Brezočnik},
   title = {Flexible job shop scheduling using zero-suppressed binary decision diagrams. Advances in production engineering and management, 13(4), pp. 373-388, 2018},
   year = 2018,
   doi = {10.14743/apem2018.4.297}
+}
+
+@online{Biddy,
+  author = {R. Meolic},
+  title = {Biddy's homepage},
+  year = 2018,
+  url = {http://biddy.meolic.com/},
+  urldate = {2019-01-15}
+}
+
+@online{BddEncyclopedia,
+  author = {R. Meolic},
+  title = {{BDD} {E}ncyclopedia},
+  year = 2018,
[…]
