+++ paper.bib
@@ -1,47 +1,47 @@
 @article{Bryant1986,
-  author = {R. E. Bryant},
-  title = {Graph-Based Algorithms for Boolean Function Manipulation. IEEE Transactions on Computers, 35(8), 677–691, 1986},
+  author = {Bryant, R. E.},
+  title = {Graph-Based Algorithms for Boolean Function Manipulation. IEEE Transactions on Computers, 35(8), pp. 677-691, 1986},
   year = 1986,
   doi = {10.1109/TC.1986.1676819}
 }
 
 @article{Knuth2009,
-  author = {D. E. Knuth},
-  title = {The Art of Computer Programming, Volume 4, Fascicle 1: Bitwise Tricks and Techniques; Binary Decision Diagrams. Addison-Wesley Professional, 2009},
+  author = {Knuth, D. E.},
+  title = {The Art of Computer Programming, {V}olume 4, {F}ascicle 1: Bitwise Tricks and Techniques; {B}inary Decision Diagrams. {A}ddison-{W}esley {P}rofessional, 2009},
   year = 2009
 }
 
 @article{Meolic2012,
-  author = {R. Meolic},
-  title = {Biddy - a Multi-platform Academic {BDD} Package. Journal of Software, 7(6), 1358–1366, 2012},
+  author = {Meolic, R.},
+  title = {Biddy - a Multi-platform Academic {BDD} Package. Journal of Software, 7(6), pp. 1358-1366, 2012},
   year = 2012,
   doi = {10.4304/jsw.7.6.1358-1366}
 }
 
 @article{Meolic2016,
-  author = {R. Meolic},
-  title = {Implementation aspects of a {BDD} package supporting general decision diagrams. {U}niversity of {M}aribor, Tech. rep., 2016},
+  author = {Meolic, R.},
+  title = {Implementation aspects of a {BDD} package supporting general decision diagrams. {U}niversity of {M}aribor, {T}ech. {r}ep., 2016},
   year = 2016,
   url = {https://dk.um.si/IzpisGradiva.php?id=68831&lang=eng},
   urldate = {2019-01-15}
 }
 
 @article{Meolic2017,
-  author = {T. van Dijk and R. Wille and R. Meolic},
+  author = {Dijk, T. van and Wille, R. and Meolic, R.},
   title = {Tagged {BDDs}: combining reduction rules from different decision diagram types. In Proc. FMCAD 2017, pp. 108-115, 2017},
   year = 2017,
   doi = {10.23919/FMCAD.2017.8102248}
 }
 
 @article{Meolic2018,
-  author = {R. Meolic and Z. Brezočnik},
+  author = {Meolic, R. and Brezočnik, Z.},
   title = {Flexible job shop scheduling using zero-suppressed binary decision diagrams. Advances in production engineering and management, 13(4), pp. 373-388, 2018},
   year = 2018,
   doi = {10.14743/apem2018.4.297}
 }
 
 @online{Biddy,
-  author = {R. Meolic},
+  author = {Meolic, R.},
   title = {Biddy's homepage},
   year = 2018,
   url = {http://biddy.meolic.com/},
@@ -49,7 +49,7 @@
 }
 
 @online{BddEncyclopedia,
-  author = {R. Meolic},
+  author = {Meolic, R.},
   title = {{BDD} {E}ncyclopedia},
   year = 2018,
   url = {http://svn.savannah.nongnu.org/viewvc/*checkout*/biddy/bddscout/ENCYCLOPEDIA/bddencyclopedia.html},
@@ -57,8 +57,8 @@
 }
 
 @article{Minato2013,
-  author = {S. Minato},
[…]
