### 2019-01-19 #1177 @brainstorm
@HaoZeke would you like to assist JOSS with a review of this software package?
