### 2019-01-20 #1189 @whedon
[REVIEW]: The Biddy BDD package
**Submitting author:** @meolic (<a href="http://orcid.org/0000-0003-3395-1227">Robert Meolic</a>)
**Repository:** <a href="https://github.com/meolic/biddy" target ="_blank">https://github.com/meolic/biddy</a>
**Version:** v1.8.2
