+++ paper.bib
@@ -43,7 +43,7 @@
 @online{Biddy,
   author = {Meolic, R.},
   title = {Biddy's homepage},
-  year = 2018,
+  year = 2019,
   url = {http://biddy.meolic.com/},
   urldate = {2019-01-15}
 }
@@ -51,7 +51,7 @@
 @online{BddEncyclopedia,
   author = {Meolic, R.},
   title = {{BDD} {E}ncyclopedia},
-  year = 2018,
+  year = 2019,
   url = {http://svn.savannah.nongnu.org/viewvc/*checkout*/biddy/bddscout/ENCYCLOPEDIA/bddencyclopedia.html},
   urldate = {2019-01-15}
 }
