+++ README.md
@@ -151,6 +151,7 @@ Biddy consists of the following core files:
 - biddyOp.c (functions for operations on BDDs)
 - biddyStat.c (functions for statistic)
 - biddyInOut.c (parsers and generators for Boolean functions)
+- biddy-example-*.* (examples)
 - package-source (script used to build distribution)
 - package-bin (script used to build distribution)
 - package-bin.bat (script used to build distribution)
@@ -176,7 +177,7 @@ Precompiled packages include dynamically linked library
 and the appropriate C header biddy.h. Currently, there are no interfaces
 for other programming languages.
 
-For linking with Biddy library you have to use:
+For linking with Biddy library you have to use (remove -lgmp if you have Biddy as a dynamic library):
 
 ~~~
 -lbiddy -lgmp
@@ -191,52 +192,26 @@ IMPORTANT:
 You should define UNIX, MACOSX, or MINGW.
 You should define USE_BIDDY if you are using Biddy via dynamic library.
 
-~~~
-/* compile with gcc -o program.exe source.c -I. -L. -lbiddy -lgmp */
-#define UNIX
+~~~c
+/* Linux + gcc: compile with gcc -DUNIX -o program.exe source.c -I. -L. -lbiddy -lgmp */
+/* WINDOWS + MINGW: compile with gcc -DMINGW -o program.exe source.c -I. -L. -lbiddy -lgmp */
+/* in this example, x1, x2, and x3 are Boolean variables */
 #include "biddy.h"
-
-#define Str2Var(x) (Biddy_GetVariable((Biddy_String)x))
-
 int main() {
   Biddy_Edge f,g,h,r;
-
-  Biddy_Init(); /* use default, i.e. ROBDD WITHOUT COMPLEMENTED EDGES */
+  Biddy_Init(); /* use default, i.e. PLAIN ROBDDS WITHOUT COMPLEMENTED EDGES */
   printf("Biddy is using %s.\n",Biddy_GetManagerName());
-
-  f = Biddy_Eval1((Biddy_String)"(OR H E L L O)"); /* PREFIX INPUT */
-  g = Biddy_Eval1((Biddy_String)"(AND W O R L D)"); /* PREFIX INPUT */
-  h = Biddy_Eval2((Biddy_String)"~(H * E * L * L)"); /* INFIX INPUT */
-
-  /* BASIC OPERATION */
-  r = Biddy_Xor(f,g);
-
-  /* REPLACE SOME VARIABLES */
-  Biddy_ResetVariablesValue();
-  Biddy_SetVariableValue(Str2Var("H"),Biddy_GetVariableEdge(Str2Var("L")));
-  Biddy_SetVariableValue(Str2Var("K"),Biddy_GetVariableEdge(Str2Var("L")));
-  Biddy_SetVariableValue(Str2Var("W"),Biddy_GetVariableEdge(Str2Var("L")));
-  r = Biddy_Replace(r);
-
-  /* SIMPLE RESTRICTION */
-  r = Biddy_Restrict(r,Str2Var("E"),FALSE);
-
-  /* COUDERT AND MADRE'S RESTRICT FUNCTION */
-  r = Biddy_Simplify(r,h);
-
-  /* SOME STATS */
-  printf("Function r depends on %u variables.\n",Biddy_DependentVariableNumber(r,FALSE));
-  printf("Function r has %.0f minterms.\n",Biddy_CountMinterms(r,0));
[…]
