### 2019-01-21 #1189 @meolic
I have improved the first example in the documentation to make the usage of the package more clear. Sorry for these last-minute changes.
