### 2019-01-21 #1189 @HaoZeke
@brainstorm, you're welcome.. I hope to have this finished in a few days.
@meolic, no problem, there may be some more changes, I will open pull requests and link to them as and when required, or discuss them with you here.

Right off the bat, could you please modify the `paper.md` so as to adhere to the following:

- [ ] Remove the historical review of package changes
- [ ] Remove or reword segments like `exciting features` and `wicked if not used carefully`

Beyond that, I would prefer a clearer statment (headings, boldface text etc) detailing the *need* for the software and *the features*.

As a style guide, perhaps [this JOSS paper](https://www.theoj.org/joss-papers/joss.01015/10.21105.joss.01015.pdf) would be useful.

Additionally, for changes to `paper.md` I would not like to open pull requests, so it might be better if I could highlight aspects which need editing, say for example if it is on [hackmd.io](https://hackmd.io/)
