+++ paper.md
@@ -20,44 +20,22 @@ bibliography: paper.bib
 
 # Summary
 
-A Binary Decision Diagram (BDD) is a data structure used in various applications
-including but not limited to the design, testing, optimization, and verification
-of digital circuits, communication protocols, and distributed systems
-[@Bryant1986] [@Knuth2009] [@Minato2013]. There exist many different types of
+A **Binary Decision Diagram** (BDD) is a data structure used in
+different areas including but not limited to the design, testing,
+optimization, and verification
+of digital circuits, communications protocols, and distributed systems
+[@Bryant1986]. There exist many different types of
 BDDs intended for different applications. The most notable types are used to
 represent Boolean functions and combination sets. For these purposes, BDDs can
 be very efficient, for example, they enable the representation and manipulation
-of sets of sparse cubes with 10^20 and more cubes [@Meolic2018].
+of set of sparse cubes with 10^47 cubes [@Minato2013].
 
-An efficient implementation of algorithms for BDDs is a rather complicated task,
-but several free BDD packages are accessible to the interested users.
-The cross-platform Biddy BDD package [@Biddy] is one of them. In fact, it is
-one of the oldest ones, with development going back to 1992 [@Meolic2012]. After
-the period of a rather slow progress, it has substantially changed throughout
-the last few years and became very competitive. Version 1.8.2 implements many
-standard functionalities, such as automatic garbage collection, complemented
-edges, and a manager. Dynamic variable ordering with a sifting algorithm and an
-exhaustive search over all the possible variable orderings are provided, too.
-Various statistics about global properties and the individual Boolean functions
-are available. Furthermore, the most distinguishing properties of Biddy are a
-refined C API and a uniform support for classical reduced ordered BDDs (ROBDDs)
-and zero-suppressed BDDs (0-sup-BDDs). At this moment, Biddy is the only
-package thoroughly supporting the tagged zero-suppressed BDDs
-[@Meolic2016] [@Meolic2017].
-Examples of BDDs of all the supported types are shown in Figure 1.
-
-![F3-88-comparison.png](./design/F3-88-comparison.png)
-Figure 1: Representation of Boolean function
-F3[088] = (NOT x1) AND x3 OR x1 AND (NOT x2) AND (NOT x3).
-From left to right there are an ROBDD, an ROBDD with complemented edges,
-a 0-sup-BDD, a 0-sup-BDD with complemented edges, and a tagged 0-sup-BDD.
-
-In BDD, every internal node contains a variable while leafs include
+In a BDD, every internal node contains a variable while leafs contain
 constants 0 and 1, respectively. For an ROBDD, each edge
 to internal node n with variable var(n), left successor else(n), and right
 successor then(n) corresponds to the Boolean function that is calculated as
 NOT(var(n)) AND else(n) OR var(n) AND then(n).
-For ROBDD in Figure 1 this is applied as:
+For the ROBDD in Figure 1 this is applied as follows:
 F3[088] =
 (NOT x1) AND (NOT(x3) AND 0 OR x3 AND 1) OR
 x1 AND (NOT(x2) AND (NOT(x3) AND 1 OR x3 AND 0) OR x2 AND 0).
@@ -66,15 +44,51 @@ leading to a leaf with constant 1 is considered to be a product of variables in
 which a negative literal is included if the path continues in the else successor
 and a positive literal is included if the path continues in the then successor.
 The resulting Boolean function is a sum of the obtained products. In this
-way, for the ROBDD in Figure 1 we obtain a minimal form
-F3[088] = (NOT x1) AND x3 OR x1 AND (NOT x2) AND (NOT x3),
+way, for the ROBDD in Figure 1 we directly obtain a minimal sum-of-products form,
 but in general, the result is not a minimal form. For the explanation of other
-types of BDDs we refer readers to the given references.
[…]
