+++ paper.bib
@@ -5,10 +5,24 @@
   doi = {10.1109/TC.1986.1676819}
 }
 
+@article{Chaki2018,
+  author = {Chaki, S. and Gurfinkel, A.},
+  title = {BDD-Based Symbolic Model Checking. In Handbook of Model Checking, Springer, pp. 219-245, 2018},
+  year = 2018,
+  doi = {10.1007/978-3-319-10575-8_8}
+}
+
 @article{Knuth2009,
   author = {Knuth, D. E.},
   title = {The Art of Computer Programming, {V}olume 4, {F}ascicle 1: Bitwise Tricks and Techniques; {B}inary Decision Diagrams. {A}ddison-{W}esley {P}rofessional, 2009},
   year = 2009
+}
+
+@article{Minato2013,
+  author = {Minato, S.},
+  title = {Techniques of {BDD}/{ZDD}: Brief History and Recent Activity. IEICE Trans. {o}n Information and Systems, {E96D}(7), pp. 1419-1429, 2013},
+  year = 2013,
+  doi = {10.1587/transinf.E96.D.1419}
 }
 
 @article{Meolic2012,
@@ -56,9 +70,3 @@
   urldate = {2019-01-15}
 }
 
-@article{Minato2013,
-  author = {Minato, S.},
-  title = {Techniques of {BDD}/{ZDD}: Brief History and Recent Activity. IEICE Trans. {o}n Information and Systems, {E96D}(7), pp. 1419-1429, 2013},
-  year = 2013,
-  doi = {10.1587/transinf.E96.D.1419}
-}
