### 2019-01-21 #1189 @meolic
@HaoZeke I have made the requested changes. Please check https://hackmd.io/Ii3fJIJmS3iplOQHSCTjHA  (i do not know how to use hackmd.io correctly).
