+++ paper.md
@@ -28,19 +28,14 @@ of digital circuits, communications protocols, and distributed systems
 BDDs intended for different applications. The most notable types are used to
 represent Boolean functions and combination sets. For these purposes, BDDs can
 be very efficient, for example, they enable the representation and manipulation
-of set of sparse cubes with 10^47 cubes [@Minato2013].
+of set of sparse cubes with $10^{47}$ cubes [@Minato2013].
 
 In a BDD, every internal node contains a variable while leafs contain
-constants 0 and 1, respectively. For an ROBDD, each edge
-to internal node n with variable var(n), left successor else(n), and right
-successor then(n) corresponds to the Boolean function that is calculated as
-NOT(var(n)) AND else(n) OR var(n) AND then(n).
+constants $0$ and $1$, respectively. For a 	Reduced Ordered Binary Decision Diagram (ROBDD), each edge to internal node $n$ with variable $var(n)$, left successor $else(n)$, and right successor $then(n)$ corresponds to the Boolean function that is calculated as $NOT$ $(var(n))$ $AND$ $else(n)$ $OR$ $var(n)$ $AND$ $then(n)$.
 For the ROBDD in Figure 1 this is applied as follows:
-F3[088] =
-(NOT x1) AND (NOT(x3) AND 0 OR x3 AND 1) OR
-x1 AND (NOT(x2) AND (NOT(x3) AND 1 OR x3 AND 0) OR x2 AND 0).
+$F3[088] = (NOT x1)\ AND\ (NOT(x3)\ AND\ 0\ OR\ x3\ AND\ 1)\ OR\ x1\ AND\ (NOT(x2)\ AND\ (NOT(x3)\ AND\ 1\ OR\ x3\ AND\ 0)\ OR\ x2\ AND\ 0)$
 The same result can be obtained if every path starting in the root and
-leading to a leaf with constant 1 is considered to be a product of variables in
+leading to a leaf with constant $1$ is considered to be a product of variables in
 which a negative literal is included if the path continues in the else successor
 and a positive literal is included if the path continues in the then successor.
 The resulting Boolean function is a sum of the obtained products. In this
@@ -50,33 +45,33 @@ types of BDDs we refer to the given references.
 
 ![F3-88-comparison.png](./design/F3-88-comparison.png)
 Figure 1: Representation of Boolean function
-F3[088] = (NOT x1) AND x3 OR x1 AND (NOT x2) AND (NOT x3) with different types of BDDs.
+$F3[088] = \overline{x1}\ \&\ x3\ + x1\ \&\ \overline{x2}\ \&\ \overline{x3}$
+with different types of BDDs.
 From left to right there are an ROBDD, an ROBDD with complemented edges,
 a 0-sup-BDD, a 0-sup-BDD with complemented edges, and a tagged 0-sup-BDD.
 
 Boolean functions are primarily used in digital circuit design where
-involving 1000 and more variables is not an unusual demand. It is impossible
-to represent so large Boolean functions with vectors and similar explicit
+ $1000$ or more variables are not unusual. It is impossible
+to represent such large Boolean functions with vectors and similar explicit
 representations of the truth table. Strings are also not an option because
 they are not canonical. On the other hand, many huge Boolean
 functions of practical importance have a managable representation with BDDs.
 Nowadays, however, logic synthesis is not the only, or even not the main, application area for
 Boolean functions and BDDs. They are also used as characteristic
-functions of sets and relations, which enables encoding
-combinatorial problems and solving them
-symbolically [@Knuth2009] [@Minato2013] [@Meolic2018].
+functions of sets and relations, which enables the encoding of
+combinatorial problems and their symbolic solution [@Knuth2009] [@Minato2013] [@Meolic2018].
[…]
