### 2019-01-22 #1189 @meolic
@HaoZeke I have updated paper.md. Many thanks for your suggestions, especially for the hint about writing math formulas in the markdown format. I am ready for your comments about the source code and the supplemented documentation (for the instalation it should be enough to read README.md, the complete API is documented in User manual available from biddy.meolic.com).
