### 2019-01-30 #1189 @brainstorm
@HaoZeke Do the updates address some of the issues?
