### 2019-01-30 #1189 @meolic
I have tried to address all of them.
