### 2019-02-06 meolic/biddy#1 @HaoZeke
Version range for libgmp
Hi, as part of the (very delayed) review https://github.com/openjournals/joss-reviews/issues/1189..

Could you mention the current `libgmp` in use? With the newest on Archlinux (6.1.2) I am having some issues, I will be testing with bundled versions from Ubuntu (via docker) but would like the information to be reflected on the `README.md` if possible
