### 2019-02-07 meolic/biddy#1 @meolic
I am using libgmp from Ubuntu repository (tested on all recent Ubuntu versions).

> apt list libgmp*
libgmp10/bionic,now 2:6.1.2+dfsg-2 amd64 [installed]
libmp-dev/bionic,now 2:6.1.2+dfsg-2 amd64 [installed]
libgmpxx4ldbl/bionic,now 2:6.1.2+dfsg-2 amd64 [installed,automatic]

Robert
