### 2019-02-07 meolic/biddy#2 @HaoZeke
Contributor guidelines
Hi, as per the checklist items for https://github.com/openjournals/joss-reviews/issues/1189, one of the requirements are that guidelines for contribution must be established.

This is doubly important for Biddy, since my understanding is that the git mirror is not where pull requests or patches should be sent. Please add some details on how patches are to be contributed (by email for example).
