+++ README.md
@@ -4,7 +4,7 @@ USER MANUAL
 ### TL;DR
 --------------
 
-Biddy is a multi-platform academic Binary Decision Diagrams package.
+**Biddy** is a multi-platform academic Binary Decision Diagrams package.
 It supports plain ROBDDs, ROBDDs with complemented edges,
 0-sup-BDDs, 0-sup-BDDs with complemented edges, and
 tagged 0-sup-BDDs.
@@ -14,33 +14,31 @@ Boolean functions, combination sets, and BDDs.
 
 Biddy is a library to be included in your C and C++ projects:
 
-~~~
+~~~c
 #include "/path/to/biddy.h"
 ~~~
 
-To compile Biddy library use "make static" or "make dynamic" or "make debug".
-There is no configuration script, you should edit Makefiles to adapt system
-configuration.
+To **compile** Biddy library use "make static" or "make dynamic" or "make debug".
+There is no configuration script, you should edit **Makefiles** to adapt the system
+configuration. Alternatively, use the prepared **Visual Studio project** (VS/Biddy.sln).
 
 ~~~
 biddy> make static
 biddy> make clean
 ~~~
 
-Dependencies:
+**Dependencies** (tested on Ubuntu and Windows 10):
 
 - on GNU/Linux, you need libgmp (https://gmplib.org/)
 - on MS Windows, you need MPIR library (http://mpir.org/)
 
-Alternatively, use the prepared Visual Studio project (VS/Biddy.sln).
-
 When using Biddy on GNU/Linux, you may have to tell bash about the library:
 
-~~~
+~~~bash
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/absolute/path/to/biddy/library
 ~~~
 
-Tests:
+**Tests**:
 
 - create a static Biddy library
 - compile the Hanoi example (notes for building are given in the source code)
@@ -52,10 +50,10 @@ biddy> gcc -DUNIX -DBIDDY -o hanoi biddy-example-hanoi.c -I. -L./bin -static -lb
 biddy> ./hanoi 10
 ~~~
 
-There are two  aditional packages included into Biddy distribution:
+There are two **aditional packages** included into Biddy distribution:
 
-- bddview is a pure Tcl/Tk script for visualization of BDDs,
-- BDD Scout is a demo application demonstrating the capability of Biddy and bddview.
+- **bddview** is a pure Tcl/Tk script for visualization of BDDs,
+- **BDD Scout** is a demo application demonstrating the capability of Biddy and bddview.
 
 To create BDD Scout (please, check its dependencies):
 
@@ -65,10 +63,13 @@ biddy> cd bddscout
 bddscout> make
 ~~~
 
-Biddy is free software maintained by Robert Meolic
-(robert.meolic@um.si) at University of Maribor, Slovenia.
+Biddy is **free software** maintained by Robert Meolic (robert@meolic.com).
 
-Homepage: http://biddy.meolic.com/
+- You can fork this project on https://github.com/meolic/biddy.
+- Fell free to report requests, bugs, and other issues on GitHub.
+- **Contributions** (patches and ideas) can be send also via author's email.
+
+**Homepage**: http://biddy.meolic.com/
 
 ### 1. AN OVERVIEW
 --------------
@@ -138,6 +139,7 @@ There are two types of C functions.
 Biddy consists of the following core files:
 
 - README.md (this file)
+- paper.md (JOSS paper)
 - CHANGES (history of changes)
 - COPYING (license file)
 - VERSION (project's version)
[…]
