### 2019-02-07 meolic/biddy#2 @meolic
Thank you for this obserrvation. I have updated the documentation. It is now included on the first page that contributions are welcome via GitHub or via email.
