### 2019-02-08 #1189 @HaoZeke
@brainstorm, I have completed my review. It is compliant with the instructions, and the code is very well documented (along with its usage). There are perhaps more edge cases (on exotic machines) however, even those are easily handled.

I believe the repository and BIDDY adheres to the JOSS guidelines and also that the claims in `paper.md` are well substantiated.

@meolic thank you for the prompt resolution of all my queries, and for open-sourcing this really well documented and useful software!

I am very sorry for the long delay.
