+++ paper.bib
@@ -12,10 +12,11 @@
   doi = {10.1007/978-3-319-10575-8_8}
 }
 
-@article{Knuth2009,
+@book{Knuth2009,
   author = {Knuth, D. E.},
   title = {The Art of Computer Programming, {V}olume 4, {F}ascicle 1: Bitwise Tricks and Techniques; {B}inary Decision Diagrams. {A}ddison-{W}esley {P}rofessional, 2009},
-  year = 2009
+  year = 2009,
+  isbn = {9780321637413}
 }
 
 @article{Minato2013,
@@ -32,7 +33,7 @@
   doi = {10.4304/jsw.7.6.1358-1366}
 }
 
-@article{Meolic2016,
+@online{Meolic2016,
   author = {Meolic, R.},
   title = {Implementation aspects of a {BDD} package supporting general decision diagrams. {U}niversity of {M}aribor, {T}ech. {r}ep., 2016},
   year = 2016,
@@ -69,4 +70,3 @@
   url = {http://svn.savannah.nongnu.org/viewvc/*checkout*/biddy/bddscout/ENCYCLOPEDIA/bddencyclopedia.html},
   urldate = {2019-01-15}
 }
-
